[Refactor] Move processors directory
E.g. `NodeHarness::Runners::Checkstyle::Processor` -> `NodeHarness::Processor::Checkstyle`
