+++ lib/node_harness.rb
@@ -31,7 +31,6 @@ require "active_support/core_ext/string/filters"
 require "node_harness/location"
 require "node_harness/results"
 require "node_harness/issues"
-require "node_harness/processor"
 require "node_harness/harness"
 require "node_harness/analyzer"
 require "node_harness/trace_writer"
@@ -54,33 +53,34 @@ require "node_harness/schema/trace"
 require "node_harness/schema/runner_config"
 require "node_harness/schema/result"
 
-require "node_harness/runners/rubocop/processor"
-require "node_harness/runners/rails_best_practices/processor"
-require "node_harness/runners/reek/processor"
-require "node_harness/runners/goodcheck/processor"
-require "node_harness/runners/querly/processor"
-require "node_harness/runners/brakeman/processor"
-require "node_harness/runners/haml_lint/processor"
-require "node_harness/runners/scss_lint/processor"
-require "node_harness/runners/code_sniffer/processor"
-require "node_harness/runners/phpmd/processor"
-require "node_harness/runners/phinder/processor"
-require "node_harness/runners/checkstyle/processor"
-require "node_harness/runners/pmd_java/processor"
-require "node_harness/runners/javasee/processor"
-require "node_harness/runners/eslint/processor"
-require "node_harness/runners/coffeelint/processor"
-require "node_harness/runners/jshint/processor"
-require "node_harness/runners/stylelint/processor"
-require "node_harness/runners/tslint/processor"
-require "node_harness/runners/tyscan/processor"
-require "node_harness/runners/flake8/processor"
-require "node_harness/runners/misspell/processor"
-require "node_harness/runners/go_vet/processor"
-require "node_harness/runners/golint/processor"
-require "node_harness/runners/gometalinter/processor"
-require "node_harness/runners/swiftlint/processor"
-require "node_harness/runners/cppcheck/processor"
+require "node_harness/processor"
+require "node_harness/processor/brakeman"
+require "node_harness/processor/checkstyle"
+require "node_harness/processor/code_sniffer"
+require "node_harness/processor/coffeelint"
+require "node_harness/processor/cppcheck"
+require "node_harness/processor/eslint"
+require "node_harness/processor/flake8"
+require "node_harness/processor/go_vet"
+require "node_harness/processor/golint"
+require "node_harness/processor/gometalinter"
+require "node_harness/processor/goodcheck"
+require "node_harness/processor/haml_lint"
+require "node_harness/processor/javasee"
+require "node_harness/processor/jshint"
+require "node_harness/processor/misspell"
+require "node_harness/processor/phinder"
+require "node_harness/processor/phpmd"
+require "node_harness/processor/pmd_java"
+require "node_harness/processor/querly"
+require "node_harness/processor/rails_best_practices"
+require "node_harness/processor/reek"
+require "node_harness/processor/rubocop"
+require "node_harness/processor/scss_lint"
+require "node_harness/processor/stylelint"
+require "node_harness/processor/swiftlint"
+require "node_harness/processor/tslint"
+require "node_harness/processor/tyscan"
 
 module NodeHarness
 end
